+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 0.15alpha
 -----------------
 - added parameter ddlx_options text[] to a bunch of functions
 - bug fixes WRT dropped attributes
+- added ddlx_apropos(pattern) function to search queries (functions and views) matching a pattern
 - support for ALTER TABLE ENABLE/FORCE ROW LEVEL SECURITY
 - support for publications